Preserve a newline between adjacent strings if the source had one. This doesn't, by design, preserving indentation when it does that, but this is still an improvement in most cases.
This also forced me to flush out a bunch of hacky implicit nesting
stuff. It's more explicit about when expression nesting is increased
now which is more verbose in the visitor, but less magical.

Filled in some more tests and allow wrapping in more places as a side
effect of that.

BUG=https://github.com/dart-lang/dart_style/issues/21
R=alanknight@google.com

Review URL: https://chromiumcodereview.appspot.com//821133004