Handle backwards implications correctly.
If a rule in the suffix implies a rule in the prefix, that can place a constraint on the suffix's rule.

537 PASSED, 50 FAILED, 0 ERRORS

This is the same as the last commit, but the behavior is more correct now. (It was broken
before any time an implied rule went from a later to an earlier rule.) The new tests that are
failing are cases where an outer rule like an arg list is incorrectly forced to split by an inner one.
We don't want all rules to require that.