Don't line split incrementally.
Feeding chunks to LineSplitter eagerly added a lot of complexity to
LineWriter with little benefit. It was a big source of bugs and hard to
understand.

This regresses perf a bit but simplifies it. I think I can get the perf back
with some simple optimization.

This also fixed a subtle bug where the old code was wrong.

560 PASSED, 29 FAILED, 0 ERRORS