Start getting nested bodies indenting correctly.
This now indents all bodies, which is wrong, but it makes it possible
to indent them at all.

Regresses perf significantly because big nested functions must be split
as a unit since they indent with nesting. Once *not* nesting functions
and collections in some circumstances is supported, this should
unregress.

Also breaks flush left lines and some other stuff. Those will be fixed
in subsequent commits.