Track indentation internally using spaces everywhere.
This is mainly a clean-up patch that makes the code a little easier to
understand. Having concepts like indentation "levels" didn't really buy
much.

Worse, that concept was exposed in the public API since the leading
indentation you could pass in was in levels, not spaces. This fixes
that.

Also, it should help me get cascades working better, which was the real
motivation for this.

R=pquitslund@google.com

Review URL: https://chromiumcodereview.appspot.com//1173303003