Optimize splitting lines with many rules.
The constraint checking was quadratic in the number of rules in the
line which is bad on pathologically long lines like #456.

This gets it back down to linear. It takes that bug from 7 minutes to
about 20 seconds on my machine. The benchmark, which is closer to real
code, gets about 20% faster.

I'm still leaving #456 open because I think there's more work to be
done there, but this is a solid improvement overall.

R=kevmoo@google.com

Review URL: https://codereview.chromium.org//1418483008 .