Avoid passing an invalid file URI to analyzer on Windows.
If you use dart_style's library API and call format() without a URI, it
defaults to using "<unknown>" for the path. Analyzer now parses that as
a file URI (as opposed to just treating it like an opaque string), and
"<" isn't allowed on Windows, so the parsing fails.

Instead, just pass null along to analyzer and let it handle it.

R=jakemac@google.com

Review URL: https://codereview.chromium.org//2174753003 .