Support optional new/const. (#669)
In almost all cases, an elided "new"/"const" is parsed as a normal
method call by analyzer, so the existing formatting behavior for those
does the right thing. The only edge case is a named constructor on a
generic class. That syntax can *only* be a constructor call, so it's
parsed as an instance creation expression.

When we get to that code path, we need to not add an extra space before
the class name if there is no preceding "new" or "const" keyword.

Fix #652.