Dogfood the new "--fix" stuff.
Use "=" for named parameter default separators. Remove unnecessary "new"
and "const". Reformat with the latest rules.

This does mean that dartfmt now only runs in --preview-dart-2, but I
think we're close to the point in time where that's OK. It has been
running in the Dart SDK with "--preview-dart-2" for a while.