Hard split in interpolation (#714)
* Allow hard splits inside string interpolation.

Removing *all* splits broken several things. In particular, multiline
strings and line comments inside interpolation really do need to
maintain their splits.

This allows all hard splits to split but not soft splits. It also allows
collection literals to split. That's not ideal, but it should be a rare
case anyway. If you don't want that to split, don't have a collection
literal inside a string interpolation in a line that doesn't fit the
page width.

Fix #711.

* Add another edge case test.

* Bump version and update changelog.

* Revise README.