Restore the command line output inadvertently removed in 1.3.4.
That release contains a new separate hidden library for running the CLI.
The new library has a different command line interface, but the current
interface exposed by "dartfmt" should be unchanged. Unfortunately, I
accidentally broke that and running the formatter stopped showing the
names of files that were formatted or unchanged.

This restores that (and adds tests, which I thought already existed).