Don't unnecessarily split argument lists with block comments.
A very subtle bug in the comment handling would cause a /* */ inline
comment to force the surrounding code to full split when not needed in
some cases.

This fixes that issue. It also slightly tweaks the formatting of code
containing inline block comments in some rare cases. In practice, the
only code I see affected by that is formatter tests. In the real world,
all affected code is related to the bug itself and is fixed by this
change.

Fix #837.