Fix hang when reading from stdin.
The tests for dart_style do not seem to trigger this bug, and I'm not
sure why not. But I can repro it in dartdev, and I have a test there
that will trigger the bug. So I'm going to push out this fix, roll it
into the SDK, and land the dartdev test with that roll.