Refactor Chunk to represent the split preceding the chunk's text.
The internal representation that the formatter users is a series of
chunks of code that can't contain any internal line breaks separated by
information describing the split between each one:

        .----. .-----. .----. .-----. .----. .-----.
    ... |text| |split| |text| |split| |text| |split| ...
        '----' '-----' '----' '-----' '----' '-----'

To have a homogeneous list of objects, each text/split pair is wrapped
into a single Chunk class. For historical reasons, they were paired up
so that each  Chunk has the text and the split that follows it:

        .--------------. .--------------. .--------------.
        |    Chunk     | |    Chunk     | |    Chunk     |
        |.----. .-----.| |.----. .-----.| |.----. .-----.|
    ... ||text| |split|| ||text| |split|| ||text| |split|| ...
        |'----' '-----'| |'----' '-----'| |'----' '-----'|
        '--------------' '--------------' '--------------'

When all we had was a flat list of chunks, this mostly didn't matter.
However, in order to improve the performance of line-splitting large
expressions containing nested function expressions and collection
literals, I moved to a more tree-like representation where some Chunks
contain a list of child Chunks.

This made having each Chunk terminate with a split more annoying to work
with because now the first child Chunk doesn't actually contain the
information to describe where it begins on its own line.

In order to be able to format child blocks independently and cache them,
it makes more sense if each Chunk describes an independent line: the
leading indentation followed by the text.

I have tried this refactoring several times in the past, but it causes
an unbelievable amount of off-by-one errors and other problems. I
finally pushed through and got everything working again.

This change doesn't actually simplify the codebase very much right now.
My goal was to get the Chunk representation changed with as little
changes to the rest of the formatter as possible. My hope is that this
will enable future refactoring and simplifications.

The external behavior is almost completely unchanged. On a corpus of
2,000 Pub packages (10+MLOC), there were <100 differences compared to
the previous formatter behavior. All of those differences were actually
(very rare) bugs, which this refactoring incidentally fixes.