Remove preserveWhitespace() and related code.
There are some places where a newline or blank line in the original
source code affects the output formatting. For example, users can put a
blank line between statements if they want.

The formatter used to track that lazily like it did other split
information. It used a Whitespace enum that could be in a fully pinned
down or still partially unknown state. Then when the subsequent comment
or text was written, it would count the newlines and figure out how to
split.

Since there is already a lot of laziness and pending state in
ChunkBuilder, that gets pretty complex.

This removes all that. Instead, right at the point in SourceVisitor
where we write whitespace that depends on the input newlines, we look
ahead and eagerly decide how to handle it.