Make cost calculate following LineWriter.
Now takes into account unsplit blocks that contain split blocks as
LineWriter does. I think it's mostly correct, but I'm not entirely
sure.

00:03 +1344 -138: Some tests failed.

This looks like a regression but skimming through the failing tests, I
think a lot of ones that were failing are passing correctly and the new
failures are often incorrect tests that were coincidentally passing
with the old broken logic.