Use a more general mechanism for tracking constraints between rules. (#1122)
* Use a more general mechanism for tracking constraints between rules.

Rules were initially decoupled from each other. Then I added _implies
to handle subexpressions forcing outer expressions to split. Then I
added some object-oriented "constrain()" and "addConstrainedRules()"
API so that Rules could dynamically add other ad-hoc constraints.

At this point, it's better to just have a single general mechanism for
expressing constraints between rules that can subsume both of those.
This does that. I think the resulting code is simpler. It is also easier
to extend for other kinds of constraints in the future.

* Add a missing argument constraint.

This (interestingly) was not caught by any existing tests, but did show
up when I ran the formatter on a big corpus. Added a test to catch it
in the future.

* Apply review feedback.