Format list, map, and record patterns.
Also test list, map, set, and record constants inside patterns.

In the process of working on this, I noticed that the formatting for
collection literal type arguments isn't ideal. In rare cases, it would
split after the initial "<", which looks totally broken. I fixed that
here too. I'd normally pull that out into a separate PR but it's mixed
in with the other changes to _visitCollectionLiteral() so that would be
hard.

Also added some more tests around avoiding splitting between a named
argument and a subsequent collection literal. The code was there to
handle it but it wasn't tested well. (It's hard to come up with an
example that hits that exact code path without formatting correctly for
other reasons.)