Format switch cases that aren't valid patterns.
Fix #1164.

The solution is kind of hacky, but users will probably never run into it
and it avoids complicated the user experience of the formatter.

To get this working, I had to update to analyzer 5.5.0 because 5.4.0 had
an assert failure when it tried to parse an invalid switch case. But
5.5.0 also has a bug which is causing a couple of formatter tests to
fail: https://github.com/dart-lang/sdk/issues/51415.

I'll probably wait until there's a fix for that out before this gets
merged to master.

Analyzer 5.5.0 also changes some of the AST types. Refactored how
binary expressions and patterns are formatted to avoid copy/paste from
that change.