Add some more tests to pin down comment indentation in switches. (#1180)
I spent a bunch of time investigating whether comments in switches that
aren't inside a case body (i.e. not between the "case" line and some
statement inside the case) should be indented to align with the cases
or the bodies.

After trying a bunch of alternatives and running them on a big corpus,
I concluded that the current behavior (always align to the cases)
actually works very well in almost all real-world code. But it's not
pinned down well by tests, so this does that.