Remove support for compact switch statements. (#1191)
I'm still interested in supporting this, but it causes a lot of churn.
That means it really needs to go through the full dart style change
process. Given how many other things are in the air with Dart 3.0, it
seems like now is probably not the ideal time for that.

So, for now, I think we should keep the old formatting where case bodies
always go on their own line.