Give record expressions block-like formatting in argument lists. (#1208)
This is consistent with how other collection literals are formatted in
argument lists.

I originally made the formatting diverge deliberately because I felt
it would make it clearer when you're looking at a record versus a series
of arguments. But based on the thumbs up on:

https://github.com/dart-lang/dart_style/issues/1205

It appears that users prefer the more consistent formatting. In
retrospect, I think I do too.