Revise splitting after type annotations.
Allow splitting after a delimited type and a variable/parameter name.
Don't indent when a split is inserted here. This is consistent with,
for example, splitting between a function declaration's return type and
its name.

It allows more record and function types to be unsplit because the
variable name isn't stuck to it.

00:05 +1814 -6: Some tests failed.