Support both the current and experimental formatting styles.
Because the experimental style is a hacky prototype implemented in
place, this essentially contains a separate copy of most of the
formatter's code inside lib/src/experiment/. Any file that has any
changes in lib/src/ in the prototype has its own separate version in
that directory.

Likewise, there is a separate copy of all of the formatting tests with
the experimental style expectations in test/experiment/. A separate
test file, test/experiment_test.dart runs those using the experimental
formatting logic.

By default, the formatter still runs exactly as it did before using the
normal style rules. (The only difference is that it checks for the opt in
marker comment below.) The experimental style can be enabled by any of:

- Invoking the formatter on the command line with "-e" or
  "--experimental-style".
- Placing a comment containing anywhere in the file:
  "// DO NOT SUBMIT USE DART FORMAT EXPERIMENT"
  It must be in a line comment. This enables the experimental style for
  that file and that file only.
- Using the formatter as a library and passing `experimentalStyle: true`
  when constructing a DartFormatter.