Start building the new Piece internal representation and new style.
This gets the foundation in place for the new IR, the new visitor that
produces it, the new line splitter that consumes it, the new style, and
new tests for it.

It also adds support for formatting library, import, and export
directives to make sure everything is wired up together and working.

Existing formatting behavior is unchanged. You have to opt in to the
new stuff by passing "tall-style" as an experiment flag.

This PR doesn't support comments, but it does leave some unused code
in a few places that will be used to handle comments in future PRs.

It has many UnimplementedError throws. Those will get filled in as more
of the language is implemented in the new style.

There are also two new kinds of TODO comments:

  TODO(tall): ...

These comments describe that work that needs to be done before the new
style is fully working.

  TODO(perf): ...

These describe potential areas for optimization. Once more of the
language is supported with the new IR and I can run some larger
benchmark samples through it, I can start exploring where the actual
performance problems are. For now, I'm just leaving breadcrumbs.