Include comments and discretionary newlines in output. (#1269)
Include comments and discretionary newlines in output.

Handling comments is the next big piece of the new representation. My
goal was to come up with something simpler than the old very hard to
maintain code in ChunkBuilder.

This seems to look pretty good at least in the subset of the language
that is supported so far.