Revamp how comments are handled.
I'm working on getting argument lists formatting, including comments,
and that pushed on a bunch of weak spots in the current way comments
are handled. So I tried to clean that up first.

The general problem is comments are a sort of "fencepost" data structure
where we need to track the number of newlines between each comment,
including before the first and after the last. The previous code modeled
that by having each comment store the number of lines before it, and
then passing around a separate int for the number of newlines after it.

This commit adds a new CommentSequence class that models it directly
and has nicer operations to get the newlines before or after any given
comment. I believe it will make argument list comments easier to
implement, and easier to share code with how sequences/blocks handle
their comments.

The main downside is that I forked the SourceComment class. There's the
old one for the old style and the new one. I couldn't come up with a
better name, so the two classes have the same name. But no library
should end up importing both of them, so it's not a problem in practice,
just a little confusing.

I also cleaned up some logic in _convertComments() that was ported
directly over from the old code but isn't needed in the new system. It
used to have to do some hacky stuff like look at the raw characters
before a comment to determine whether to split before. That was because
the old IR doesn't have much context when it's handling comments. With
the new sequence/non-sequence comment approach, we know directly whether
we're at the beginning of a block/sequence.