Move handling comments in sequences into SequenceBuilder.
I had an idea that it would make sense to consolidate all comment
handling in CommentWriter. But with the forthcoming argument list
support, it looks like there will be three separate styles of comments.
At that point, it probably makes more sense to have the comment handling
for each construct with that construct.

So I moved the code in SequencePiece that incrementally builds it and
the code in CommentWriter that handles sequence comments into a new
SequenceBuilder class.

I think this is overall a cleaner design. I like that SequencePiece is
now immutable.