Allow non-contiguous state values.
Requiring the states to be a zero-based and contiguous makes it harder
for some pieces to support different sets of different states based on
what they contain. For example, an ImportPiece only supports states for
splitting the names inside a single combinator when there is more than
one combinator.

To model that with contiguous states requires the piece to carefully
track how state indexes are shifted or splitting pieces into different
classes to handle the different sets of states.

Instead, with this change, we simply allow the piece to return a list
of states that may be non-contiguous. They can then omit states that
don't make sense given the piece's contents. This makes ImportPiece
simpler and I think will be useful for other pieces that have a variety
of ways they can split.

Also wrap state values in a class to make things a little clearer and
more type safe.