Format list literals (and some other things). (#1281)
Format list literals (and some other things).

I'm trying to keep these PRs mostly well-contained and not just adding a
random pile of formatting. But I'm also trying to make sure I don't drop
stuff on the floor. For this one, I filled in a few more things that
lists touch on:

- Handle type argument lists in lists and function calls.
- Support block-like formatting for right-hand sides of `=`.
- Format parenthesized expressions.
- Format named type annotations with type arguments and nullable types.
- Add support for states with different costs so we can prioritize splitting choices.

Co-authored-by: Nate Bosch <nbosch@google.com>