Format map and set literals. (#1283)
I'm formatting map entries the same way that assignments and variable
declarations are handled, so that we get nice "block-like" formatting
for delimited expressions, as in:

```
var map = {
  list: [
    1,
    2,
  ]
};
```

While I was at it, I also updated named arguments to do the same thing:

```
function(
  list: [
    1,
    2,
  ]
);
```

Also handle the "const" keyword on lists. I missed that when adding
initial support for lists.