Format while, do-while, break, and continue statements.
Reuses the existing formatting logic for if statements to format while
statements. (A goal of the rewrite is to reuse as much formatting code
as possible across different syntactic elements so that the architecture
of the formatter itself ensures consistent formatting.)

Do-while loops are kind of weird, so they get their own handling.