Support eagerly pinning a piece to a specific state.
The BlockPiece had some bespoke support for being in an "always split"
configuration that overrode the two different ways it could be
formatted. While working on conditional expressions, I found myself
needing something similar to InfixPiece.

Instead of adding similar code to that and likely other pieces in the
future, I figured it made more sense to bake support for locking a
piece into a specific state directly in the Piece base class.