Format for statements.
This handles:

```
for (initializer; condition; increment) ...
for (var variable; condition; increment) ...
```

It doesn't add support for:

```
// No patterns yet:
for (var [pattern]; condition; increment) ...

// No for-each yet:
for (existingVariable in iterable) ...
for (var newVariable in iterable) ...
// No for-each or patterns yet:
for (var [pattern] in iterable) ...
```

This one was a lot of work. There were almost no tests for for loops
already and the style approach I took here is different from the old
style. Since the clauses are formatted more like a list now, I wanted
to reuse DelimitedListBuilder but it took some work to get that playing
nice with clauses that aren't comma separated.

That's what the previous refactoring PR is to enable.

After this, supporting for-each and for-with-patterns should be (I
hope!) relatively straightforward.