Refactor ListPiece and DelimitedListBuilder. (#1295)
Refactor ListPiece and DelimitedListBuilder.

These classes started out just supporting argument lists, then grew to
collection literals, then parameter lists, then switch expressions,
and soon for loop headers.

Each time, I need to tweak how it works a little. (But not enough to
not use it completely, since the comment handling is complex and is
the same across all of those constructors.)

There was getting to be an unwieldy set of options being passed around.
Cleaned that up by defining a separate ListStyle that contains those
options and can be shuttled directly from DelimitedListBuilder to
ListPiece.

I made the before and after bracket pieces optional since I'll be using ListPiece in a future PR for the updaters list in a for loop, which isn't delimited.

Did a bunch of other minor clean-up.

Also, many of the docs were out of date. Fixed those.