Update test running code in example script.
The example script in example/format.dart has long been my sandbox
script for debugging the formatter. One thing it has is a little helper
function to run one of the format tests.

That helper code was copy/pasted from an ancient version of the test
running code. Cleaned it up here to actually use the TestFile class
now that there is some abstraction for running tests.

Also added support to example/format.dart to handle selections in tests.

(I'm starting to work on handling selections in the new formatter.)