Make a new test folder for variable tests.
Variables (and soon, functions) are sort of weird because they can
occur at the top level and as statements. It's not clear where they
should go. Before this PR, they were in both top_level/ and statement/.
This moves them to their own test directory.

I'll be doing the same thing for functions too.