Get selections running through the new back end. (#1300)
Get selections running through the new back end.

This mostly ports similar code directly over from the old back end,
including a couple of quirks in its behavior around whitespace. I
considered simplifying that, but I think it's better to have fewer
changes during this already huge migration.

Because the new back end adds and removes trailing commas, it treats
all commas as whitespace. When a selection is in whitespace, it will be
snapped to the nearest non-whitespace code. These two interact in that
a selection right around a comma may snap over it because the comma is
considered whitespace.

I thought about making it smarter but I didn't want to rathole too long.
Instead, I left a comment in CodeWriter.write() with some context and
we can revisit it when the overall rewrite is further along. We may end
up just deleting that comment later if we decide the behavior here is
good enough (I think it is), but this way to remember to think about it.