Format function declarations and expressions.
This covers:

- Top-level function declarations
- Local function declaration statements
- Anonymous function expressions

It handles all three kinds of bodies: `;`, `{ ... }`, and `=> ... ;`.

Most of the implementation work for functions is in:

- Their parameter lists, which is most already handled by the existing
  support (and tests) for function type annotations.

- Their `{ ... }` bodies, which are just block statements and are
  already implemented and tested.

- Their `=> ...` bodies, which are formatted like assignments so need
  little additional implementation.

This PR mostly just stitches that together, adds tests, and covers some
of the bits that can't appear in those existing constructs:

- The `external` modifier on function declarations.

- Function names.

- Default values.

- The `var` and `final` modifiers on parameters.

This PR does not do getters and setters. That will come later.